# AI_Middle_Office/openapi/bid-assessment-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "旗胜投标机会研判 Agent API",
    "version": "1.0.0-draft.2",
    "description": "报价资料研判 Agent 机器合同；已冻结 Phase 2 文档处理与 Phase 3A—3D Run Bootstrap、Planner、Task Runtime、API-40—API-43 边界，但不代表相关运行时已经上线。"
  },
  "servers": [{"url": "/api/v1"}],
  "security": [{"bearerAuth": []}],
...
  "paths": {
...
    "/bid-assessments/{assessment_id}/reports": {
      "parameters": [{"$ref": "#/components/parameters/AssessmentId"}],
      "get": {
        "x-api-id": "API-60",
        "operationId": "listBidReports",
        "tags": ["reports"],
        "parameters": [{"$ref": "#/components/parameters/Page"}, {"$ref": "#/components/parameters/PageSize"}],
        "responses": {"200": {"$ref": "#/components/responses/GenericOk"}, "default": {"$ref": "#/components/responses/Error"}}
      }
    },
    "/bid-reports/{report_id}": {
      "parameters": [{"$ref": "#/components/parameters/ReportId"}],
      "get": {
        "x-api-id": "API-61",
        "operationId": "getBidReport",
        "tags": ["reports"],
        "responses": {"200": {"$ref": "#/components/responses/ReportOk"}, "304": {"description": "ETag 未变化"}, "default": {"$ref": "#/components/responses/Error"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "parameters": {
...
      "Page": {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "default": 1}},
      "PageSize": {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}},
...
      "AssessmentId": {"name": "assessment_id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 80}},
...
      "ReportId": {"name": "report_id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 80}},
...
    },
    "schemas": {
...
      "SuccessEnvelope": {"$ref": "../schemas/bid_assessment/v1/contracts.schema.json#/$defs/SuccessEnvelope"},
...
      "ReportResponse": {"$ref": "../schemas/bid_assessment/v1/contracts.schema.json#/$defs/ReportResponse"},
      "ErrorEnvelope": {"$ref": "../schemas/bid_assessment/v1/contracts.schema.json#/$defs/ErrorEnvelope"}
    },
    "responses": {
      "GenericOk": {"description": "成功", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuccessEnvelope"}}}},
...
      "ReportOk": {"description": "已发布报告", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReportResponse"}}}},
      "Error": {"description": "稳定业务错误", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}
    }
  }
}